chore: separate the new TS files from actual data files
