+++ src/mappings/fonts.config.ts
